pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "server-manager"
dynamic = ["version"]
description = "Server Manager"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "NS", email = "nathanswanson370@gmail.com" }]
dependencies = [
    "click",
    "fastapi[standard]",
    "python-multipart",
    "bcrypt",
    "cryptography",
    "aiodocker",
    "pyJWT",
    "sqlmodel",
    "psycopg2-binary",
    "websockets",
    "miniupnpc",
    "python-socketio",
    "colorama",
    "requests",
    "strawberry-graphql[fastapi]",
]

[project.scripts]
server_manager = "server_manager.cli:server_manager"

[tool.hatch.version]
path = "src/server_manager/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/server_manager"]
include = ["src/*.py", "src/resources/*"]

[tool.hatch.envs.hatch-static-analysis]
dependencies = ["ruff>=0.12.8"]
config-path = "ruff.toml"

# dev
[tool.hatch.envs.default]
dependencies = [
    "debugpy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "requests-mock",
    "pytest-mock",
    "pytest-xdist",
]
include = ["src/*.py", "src/resources/*"]

[tool.hatch.envs.default.scripts]
debug = "python -Xfrozen_modules=off -m debugpy --listen 0.0.0.0:5678  -m uvicorn src.server_manager.webservice.webservice:app --reload --host 0.0.0.0 --port 8000"

# test
[tool.hatch.envs.hatch-test]
dependencies = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xdist",
    "requests-mock",
]

[tool.pytest.ini_options]
addopts = "-n auto --cov=./src/server_manager/ --cov-branch --cov-config=./pyproject.toml --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/server_manager", "tests"]
omit = [
    "src/server_manager/__about__.py",
    "src/server_manager/__main__.py",
    "src/server_manager/cli/__init__.py",
]
